+++ package.json
@@ -27,7 +27,7 @@
     "lodash": "^4.17.15",
     "merge-images": "^2.0.0",
     "path": "^0.12.7",
-    "qrcode.react": "^1.0.0",
+    "qrcode.react": "^3.1.0",
     "react": "^18.3.1",
     "react-beautiful-dnd": "^13.1.1",
     "react-dom": "^18.3.1",
